Bag of words in NLP
The file can be used to create a bag of words model from a paragraph. All the sentences in the paragraph are compared to each other and at the end, we get a Numpy array which gives the bag of words. The number of words can be changed by changing the parameter of the heapq. It is now set to 100.